+++ src/components/CensusView/index.ts
@@ -1,0 +1,3 @@
+export { CensusView } from "./CensusView";
+export type { CensusViewProps } from "./censusModel";
+export * from "./censusModel";